requirements.txt
ajax==3.0.0
celery==5.2.7
Django==4.1.7
django-celery-beat==2.5.0
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
Pillow==9.4.0
psycopg2==2.9.5
redis==4.5.3
